+++ pkgs/firehose/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.3.32
 
-- Add PR Health checks for `DO_NOT_SUBMIT` strings.
+- Add PR Health checks for `DO_NOT${'_'}SUBMIT` strings.
 
 ## 0.3.31